+++ partiql-lang/src/test/kotlin/org/partiql/lang/planner/transforms/AstToLogicalVisitorTransformTests.kt
@@ -780,7 +780,8 @@ class AstToLogicalVisitorTransformTests {
                                     ),
                                     lit(ionInt(2))
                                 )
-                            )
+                            ),
+                            rowAlias = varDecl(AstToLogicalVisitorTransform.EXCLUDED)
                         ),
                         bindingsToValues(
                             struct(structFields(id("x", caseInsensitive(), unqualified()))),
@@ -822,7 +823,38 @@ class AstToLogicalVisitorTransformTests {
                                         ),
                                         lit(ionInt(2))
                                     )
+                                ),
+                                rowAlias = varDecl(AstToLogicalVisitorTransform.EXCLUDED)
+                            ),
+                            bag(
+                                struct(
+                                    structField(lit(ionString("id")), lit(ionInt(1))),
+                                    structField(lit(ionString("name")), lit(ionString("bob")))
                                 )
+                            )
+                        )
+                    }
+                }
+            ),
+            // Testing using excluded non-reserved keyword in condition
+            TestCase(
+                "INSERT INTO foo AS f <<{'id': 1, 'name':'bob'}>> ON CONFLICT DO REPLACE EXCLUDED WHERE excluded.id > 2",
+                PartiqlLogical.build {
+                    PartiqlLogical.build {
+                        dml(
+                            identifier("foo", caseInsensitive()),
+                            dmlReplace(
+                                varDecl("f"),
+                                condition = gt(
+                                    listOf(
+                                        path(
+                                            id("excluded", caseInsensitive(), unqualified()),
+                                            listOf(pathExpr(lit(ionString("id")), caseInsensitive()))
+                                        ),
+                                        lit(ionInt(2))
+                                    )
+                                ),
+                                rowAlias = varDecl(AstToLogicalVisitorTransform.EXCLUDED)
                             ),
                             bag(
                                 struct(
@@ -880,7 +912,38 @@ class AstToLogicalVisitorTransformTests {
                                         ),
                                         lit(ionInt(2))
                                     )
+                                ),
+                                rowAlias = varDecl(AstToLogicalVisitorTransform.EXCLUDED)
+                            ),
+                            bag(
+                                struct(
+                                    structField(lit(ionString("id")), lit(ionInt(1))),
+                                    structField(lit(ionString("name")), lit(ionString("bob")))
                                 )
+                            )
+                        )
+                    }
+                }
+            ),
+            // Testing using excluded non-reserved keyword in condition
+            TestCase(
+                "INSERT INTO foo AS f <<{'id': 1, 'name':'bob'}>> ON CONFLICT DO UPDATE EXCLUDED WHERE excluded.id > 2",
+                PartiqlLogical.build {
+                    PartiqlLogical.build {
+                        dml(
+                            identifier("foo", caseInsensitive()),
+                            dmlUpdate(
+                                varDecl("f"),
+                                condition = gt(
+                                    listOf(
+                                        path(
+                                            id("excluded", caseInsensitive(), unqualified()),
+                                            listOf(pathExpr(lit(ionString("id")), caseInsensitive()))
+                                        ),
+                                        lit(ionInt(2))
+                                    )
+                                ),
+                                rowAlias = varDecl(AstToLogicalVisitorTransform.EXCLUDED)
                             ),
                             bag(
                                 struct(