+++ partiql-lang/src/test/kotlin/org/partiql/lang/planner/transforms/LogicalToLogicalResolvedVisitorTransformTests.kt
@@ -867,24 +867,40 @@ class LogicalToLogicalResolvedVisitorTransformTests {
     @ArgumentsSource(DmlStatements::class)
     fun `dml statements`(tc: TestCase) = runTestCase(tc)
     class DmlStatements : ArgumentsProviderBase() {
+        val EXCLUDED = AstToLogicalVisitorTransform.EXCLUDED
         override fun getParameters() = listOf(
             TestCase(
                 "INSERT INTO foo << {'a': 1} >> ON CONFLICT DO REPLACE EXCLUDED WHERE foo.id > 2",
                 Expectation.Success(
                     ResolvedId(1, 70) { localId(0) },
-                ).withLocals(localVariable("foo", 0))
+                ).withLocals(localVariable("foo", 0), localVariable(EXCLUDED, 1))
             ),
             TestCase(
                 "INSERT INTO foo AS f << {'a': 1} >> ON CONFLICT DO REPLACE EXCLUDED WHERE f.id > 2",
                 Expectation.Success(
                     ResolvedId(1, 75) { localId(0) },
-                ).withLocals(localVariable("f", 0))
+                ).withLocals(localVariable("f", 0), localVariable(EXCLUDED, 1))
+            ),
+            // Tests that we can use non-reserved excluded in on-conflict condition
+            TestCase(
+                "INSERT INTO foo AS f << {'a': 1} >> ON CONFLICT DO REPLACE EXCLUDED WHERE excluded.id > f.id",
+                Expectation.Success(
+                    ResolvedId(1, 89) { localId(0) },
+                    ResolvedId(1, 75) { localId(1) },
+                ).withLocals(localVariable("f", 0), localVariable(EXCLUDED, 1))
+            ),
+            // Tests that we cannot use unresolved non-reserved EXCLUDED. See the insert rows value. No self-referencing.
+            TestCase(
+                "INSERT INTO foo AS f << {'a': excluded} >> ON CONFLICT DO REPLACE EXCLUDED WHERE f.id > 2",
+                Expectation.Problems(
+                    problem(1, 31, PlanningProblemDetails.UndefinedVariable("excluded", false))
+                )
             ),
             TestCase(
                 "INSERT INTO foo AS f << {'a': 1} >> ON CONFLICT DO REPLACE EXCLUDED WHERE foo.id > 2",
                 Expectation.Success(
                     ResolvedId(1, 75) { globalId("fake_uid_for_foo") },
-                ).withLocals(localVariable("f", 0))
+                ).withLocals(localVariable("f", 0), localVariable(EXCLUDED, 1))
             ),
             TestCase(
                 "INSERT INTO foo << {'a': 1} >> ON CONFLICT DO REPLACE EXCLUDED WHERE f.id > 2",
@@ -896,13 +912,28 @@ class LogicalToLogicalResolvedVisitorTransformTests {
                 "INSERT INTO foo << {'a': 1} >> ON CONFLICT DO UPDATE EXCLUDED WHERE foo.id > 2",
                 Expectation.Success(
                     ResolvedId(1, 69) { localId(0) },
-                ).withLocals(localVariable("foo", 0))
+                ).withLocals(localVariable("foo", 0), localVariable(EXCLUDED, 1))
             ),
             TestCase(
                 "INSERT INTO foo AS f << {'a': 1} >> ON CONFLICT DO UPDATE EXCLUDED WHERE f.id > 2",
                 Expectation.Success(
                     ResolvedId(1, 74) { localId(0) },
-                ).withLocals(localVariable("f", 0))
+                ).withLocals(localVariable("f", 0), localVariable(EXCLUDED, 1))
+            ),
+            // Tests that we can use non-reserved excluded in on-conflict condition
+            TestCase(
+                "INSERT INTO foo AS f << {'a': 1} >> ON CONFLICT DO UPDATE EXCLUDED WHERE excluded.id > f.id",
+                Expectation.Success(
+                    ResolvedId(1, 88) { localId(0) },
+                    ResolvedId(1, 74) { localId(1) },
+                ).withLocals(localVariable("f", 0), localVariable(EXCLUDED, 1))
+            ),
+            // Tests that we cannot use unresolved non-reserved EXCLUDED. See the insert rows value. No self-referencing.
+            TestCase(
+                "INSERT INTO foo AS f << {'a': excluded} >> ON CONFLICT DO UPDATE EXCLUDED WHERE f.id > 2",
+                Expectation.Problems(
+                    problem(1, 31, PlanningProblemDetails.UndefinedVariable("excluded", false))
+                )
             ),
             TestCase(
                 "INSERT INTO foo << {'a': 1} >> ON CONFLICT DO UPDATE EXCLUDED WHERE f.id > 2",
@@ -914,7 +945,7 @@ class LogicalToLogicalResolvedVisitorTransformTests {
                 "INSERT INTO foo AS f << {'a': 1} >> ON CONFLICT DO UPDATE EXCLUDED WHERE foo.id > 2",
                 Expectation.Success(
                     ResolvedId(1, 74) { globalId("fake_uid_for_foo") },
-                ).withLocals(localVariable("f", 0))
+                ).withLocals(localVariable("f", 0), localVariable(EXCLUDED, 1))
             ),
         )
     }